refactor: extract ViewInvalidationService from MutationManager
Move view lifecycle logic out of MutationManager into a dedicated
ViewInvalidationService. The extracted logic is pure graph/view
orchestration (redirect writes to identity views' source schemas,
invalidate dependent view caches, topologically order cascade views,
spawn background precomputation) — it has nothing to do with mutation
execution per se.

MutationManager now owns atoms/molecules/idempotency/events, and
delegates view lifecycle events to ViewInvalidationService. Shrinks
mutation_manager.rs from ~1316 to ~884 LOC.

Pure refactor — no behavior changes. All view integration tests pass.
